Pull SettingItem into list components and reuse in SettingView and BitcoinUnitView
