+++ src/component/list.js
@@ -7,7 +7,9 @@ import {
   StyleSheet,
 } from 'react-native';
 import PropTypes from 'prop-types';
-import { color } from './style';
+import Text from './text';
+import Icon from './icon';
+import { color, font } from './style';
 
 //
 // List Content
@@ -116,5 +118,73 @@ export const ListHeader = ({ style, children }) => (
 
 ListHeader.propTypes = {
   children: PropTypes.node,
+  style: ViewPropTypes.style,
+};
+
+//
+// Setting Item
+//
+
+const iStyles = StyleSheet.create({
+  item: {
+    height: 60,
+  },
+  name: {
+    flex: 1,
+    color: color.grey,
+    fontSize: font.sizeSub,
+  },
+  lbl: {
+    fontSize: font.sizeS,
+    color: color.greyListLabel,
+    opacity: 0.74,
+  },
+  frwd: {
+    height: 15 * 0.9,
+    width: 9 * 0.9,
+    marginLeft: 20,
+  },
+});
+
+export const SettingItem = ({ name, onSelect, label, arrow, children }) => (
+  <ListItem style={iStyles.item} onSelect={onSelect}>
+    <Text style={iStyles.name}>{name}</Text>
+    {label ? <Text style={iStyles.lbl}>{label}</Text> : null}
+    {children}
+    {arrow ? <Icon image="forward" style={iStyles.frwd} /> : null}
+  </ListItem>
+);
+
+SettingItem.propTypes = {
+  name: PropTypes.string.isRequired,
+  onSelect: PropTypes.func.isRequired,
+  label: PropTypes.string,
+  arrow: PropTypes.bool,
+  children: PropTypes.node,
+};
+
+//
+// Setting Header
+//
+
+const hStyles = StyleSheet.create({
+  header: {
+    height: 20,
+  },
+  txt: {
+    fontFamily: 'OpenSans SemiBold',
+    color: color.greyListHeader,
+    fontSize: font.sizeXS,
+  },
+});
+
+export const SettingHeader = ({ name, style }) => (
+  <ListHeader style={[hStyles.header, style]}>
+    <Text style={[iStyles.i, hStyles.txt]}>{name}</Text>
+  </ListHeader>
+);
+
+SettingHeader.propTypes = {
+  name: PropTypes.string.isRequired,
   style: ViewPropTypes.style,
 };
